+++ embedded-graphics/src/lib.rs
@@ -170,7 +170,7 @@
 //! [`Point`]: ./geometry/struct.Point.html
 //! [`Size`]: ./geometry/struct.Size.html
 //! [`Font6x8`]: ./fonts/struct.Font6x8.html
-//! [`DrawTarget`]: ./trait.DrawTarget.html
+//! [`DrawTarget`]: ./draw_target/trait.DrawTarget.html
 //! [`Drawable`]: ./drawable/trait.Drawable.html
 
 #![doc(
@@ -191,6 +191,7 @@
 extern crate nalgebra;
 
 mod check_readme;
+mod draw_target;
 pub mod drawable;
 pub mod fonts;
 pub mod geometry;
@@ -202,283 +203,4 @@ pub mod primitives;
 pub mod style;
 pub mod transform;
 
-use crate::{
-    drawable::Drawable,
-    geometry::{Dimensions, Point, Size},
-    pixelcolor::PixelColor,
-    primitives::Primitive,
-    style::{PrimitiveStyle, Styled},
-};
-
-/// Defines a display that can be used to render [`Drawable`] objects.
-///
-/// To use this crate in a driver, `DrawTarget` must be implemented. This trait defines how a
-/// display draws pixels, and optionally provides a way to define accelerated drawing methods for
-/// graphical primitives such as lines, rectangles, triangles, and circles.
-///
-/// Once a `DrawTarget` is defined, it can be used to render [`Drawable`]s. Note that any iterator
-/// over [`Pixel`]s has a default implementation for the [`Drawable`] trait. See the [`Drawable`]
-/// trait documentation for more details.
-///
-/// [`Drawable`]: ./drawable/trait.Drawable.html
-/// [`Pixel`]: ./drawable/struct.Pixel.html
-///
-/// Here's an example for an imaginary display that has a 64x64px framebuffer of 8 bit values that
-/// communicates over a (simplified) SPI interface:
-///
-/// ```rust
-/// use embedded_graphics::{
-///     drawable::Pixel,
-///     egcircle,
-///     geometry::Size,
-///     pixelcolor::{Gray8, GrayColor},
-///     prelude::*,
-///     primitive_style, DrawTarget,
-/// };
-///
-/// # struct SPI1;
-/// #
-/// # impl SPI1 {
-/// #     pub fn send_bytes(&self, buf: &[u8]) -> Result<(), ()> {
-/// #         Ok(())
-/// #     }
-/// # }
-/// #
-/// /// A fake display 64px x 64px where each pixel is stored as a single `u8`
-/// struct ExampleDisplay {
-///     framebuffer: [u8; 64 * 64],
-///     iface: SPI1,
-/// }
-///
-/// impl ExampleDisplay {
-///     /// Send buffer to the display
-///     pub fn flush(&self) -> Result<(), ()> {
-///         self.iface.send_bytes(&self.framebuffer)
-///     }
-/// }
-///
-/// impl DrawTarget<Gray8> for ExampleDisplay {
-///     /// Draw a `Pixel` that has a color defined as `Gray8`.
-///     fn draw_pixel(&mut self, pixel: Pixel<Gray8>) {
-///         let Pixel(coord, color) = pixel;
-///         // Place an (x, y) pixel at the right index in the framebuffer
-///         let index = coord.x + coord.y * 64;
-///         self.framebuffer[index as usize] = color.luma();
-///     }
-///
-///     fn size(&self) -> Size {
-///         Size::new(64, 64)
-///     }
-/// }
-///
-/// fn main() {
-///     let mut display = ExampleDisplay {
-///         framebuffer: [0; 4096],
-///         iface: SPI1,
-///     };
-///
-///     // Draw a circle centered around `(32, 32)` with a radius of `10` and a white stroke
-///     let circle = egcircle!(
-///         center = (32, 32),
-///         radius = 10,
-///         style = primitive_style!(stroke_color = Gray8::WHITE)
-///     );
-///     circle.draw(&mut display);
-///
-///     // Update the display
-///     display.flush().expect("Failed to send data to display");
-/// }
-/// ```
-///
-/// ## Hardware Acceleration
-///
-/// In addition to defining [`draw_pixel`], an implementation of [`DrawTarget`] can also provide
-/// alternative implementations of drawing methods for graphical primitives. Here is an example of
-/// how a display with accelerated methods can implement [`DrawTarget`]:
-///
-/// [`draw_pixel`]: ./trait.DrawTarget.html#method.draw_pixel
-/// [`DrawTarget`]: ./trait.DrawTarget.html
-///
-/// ```rust
-/// # use embedded_graphics::prelude::*;
-/// # use embedded_graphics::DrawTarget;
-/// # use embedded_graphics::{egrectangle, primitive_style};
-/// # use embedded_graphics::primitives::rectangle::Rectangle;
-/// # use embedded_graphics::pixelcolor::{Gray8, GrayColor};
-/// # use embedded_graphics::drawable::Pixel;
-/// # use embedded_graphics::style::{PrimitiveStyle, Styled};
-/// #
-/// # struct SPI1;
-/// #
-/// # impl SPI1 {
-/// #     pub fn send_bytes(&self, buf: &[u8]) -> Result<(), ()> {
-/// #         Ok(())
-/// #     }
-/// # }
-/// #
-/// /// A fake display 64px x 64px where each pixel is stored as a single `u8`
-/// struct FastExampleDisplay {
-///     framebuffer: [u8; 64 * 64],
-///     iface: SPI1,
-/// }
-///
-/// impl FastExampleDisplay {
-///     /// Send buffer to the display
-///     pub fn flush(&self) -> Result<(), ()> {
-///         self.iface.send_bytes(&self.framebuffer)
-///     }
-///
-///     /// A HW-accelerated method for drawing rectangles
-///     pub fn fast_rectangle(&self, rect: &Styled<Rectangle, PrimitiveStyle<Gray8>>) {
-///         // Does some speedy drawing
-///     }
-/// }
-///
-/// impl DrawTarget<Gray8> for FastExampleDisplay {
-///     /// Draw a `pixel` that has a color defined as `Gray8`
-///     fn draw_pixel(&mut self, pixel: Pixel<Gray8>) {
-///         let Pixel(coord, color) = pixel;
-///         // Place an (x, y) pixel at the right index in the framebuffer
-///         let index = coord.x + coord.y * 64;
-///         self.framebuffer[index as usize] = color.luma();
-///     }
-///
-///     fn size(&self) -> Size {
-///         Size::new(64, 64)
-///     }
-///
-///     /// Use the accelerated method when drawing rectangles
-///     fn draw_rectangle(&mut self, item: &Styled<Rectangle, PrimitiveStyle<Gray8>>) {
-///         self.fast_rectangle(item);
-///     }
-/// }
-///
-/// fn main() {
-///     let mut display = FastExampleDisplay {
-///         framebuffer: [0; 4096],
-///         iface: SPI1,
-///     };
-///
-///     // Draw a rectangle from (10, 20) to (30, 40) with a white stroke
-///     let rect = egrectangle!(
-///         top_left = (10, 20),
-///         bottom_right = (30, 40),
-///         style = primitive_style!(stroke_color = Gray8::WHITE)
-///     );
-///     rect.draw(&mut display); // Uses the accelerated draw_rectangle function
-///
-///     // Update the display
-///     display.flush().expect("Failed to send data to display");
-/// }
-/// ```
-pub trait DrawTarget<C>
-where
-    C: PixelColor,
-{
-    /// Draws a pixel on the display.
-    ///
-    /// Note that some displays require a "flush" operation
-    /// to actually write changes to the framebuffer.
-    fn draw_pixel(&mut self, item: drawable::Pixel<C>);
-
-    /// Draws an object from an iterator over its pixels.
-    fn draw_iter<T>(&mut self, item: T)
-    where
-        T: IntoIterator<Item = drawable::Pixel<C>>,
-    {
-        for pixel in item {
-            self.draw_pixel(pixel);
-        }
-    }
-
-    /// Returns the dimensions of the `DrawTarget` in pixels.
-    fn size(&self) -> Size;
-
-    /// Clears the display with the supplied color.
-    ///
-    /// This default implementation should be replaced if the implementing driver provides an
-    /// accelerated clearing method.
-    fn clear(&mut self, color: C)
-    where
-        Self: Sized,
-    {
-        primitives::Rectangle::new(Point::zero(), Point::zero() + self.size())
-            .into_styled(PrimitiveStyle::with_fill(color))
-            .draw(self);
-    }
-
-    /// Flushes changes to the framebuffer.
-    ///
-    /// Note that some displays operate in "immediate mode", which does not require any flushing.
-    /// Because of this, the default implementation of this method is a noop. If the implementing
-    /// display requires flushing, this method should be overriden in the trait impl.
-    fn flush(&mut self) {}
-
-    /// Draws a line primitive.
-    ///
-    /// This default trait method should be overridden if a display provides hardware-accelerated
-    /// methods for drawing lines.
-    ///
-    /// # Caution
-    ///
-    /// This method should not be called directly from application code. It is used to define the
-    /// internals of the [`draw`] method used for the [`Line`] primitive. To draw a line, call
-    /// [`draw`] on a [`Line`] primitive object.
-    ///
-    /// [`Line`]: ./primitives/line/struct.Line.html
-    /// [`draw`]: ./trait.DrawTarget.html#method.draw
-    fn draw_line(&mut self, item: &Styled<primitives::Line, PrimitiveStyle<C>>) {
-        self.draw_iter(item);
-    }
-
-    /// Draws a triangle primitive.
-    ///
-    /// This default trait method should be overridden if a display provides hardware-accelerated
-    /// methods for drawing triangles.
-    ///
-    /// # Caution
-    ///
-    /// This method should not be called directly from application code. It is used to define the
-    /// internals of the [`draw`] method used for the [`Triangle`] primitive. To draw a triangle, call
-    /// [`draw`] on a [`Triangle`] primitive object.
-    ///
-    /// [`Triangle`]: ./primitives/triangle/struct.Triangle.html
-    /// [`draw`]: ./trait.DrawTarget.html#method.draw
-    fn draw_triangle(&mut self, item: &Styled<primitives::Triangle, PrimitiveStyle<C>>) {
-        self.draw_iter(item);
-    }
-
-    /// Draws a rectangle primitive.
-    ///
-    /// This default trait method should be overridden if a display provides hardware-accelerated
-    /// methods for drawing rectangle.
-    ///
-    /// # Caution
-    ///
-    /// This method should not be called directly from application code. It is used to define the
-    /// internals of the [`draw`] method used for the [`Rectangle`] primitive. To draw a rectangle, call
-    /// [`draw`] on a [`Rectangle`] primitive object.
-    ///
-    /// [`Rectangle`]: ./primitives/rectangle/struct.Rectangle.html
-    /// [`draw`]: ./trait.DrawTarget.html#method.draw
-    fn draw_rectangle(&mut self, item: &Styled<primitives::Rectangle, PrimitiveStyle<C>>) {
-        self.draw_iter(item);
-    }
-
-    /// Draws a circle primitive.
-    ///
-    /// This default trait method should be overridden if a display provides hardware-accelerated
-    /// methods for drawing circles.
-    ///
-    /// # Caution
-    ///
-    /// This method should not be called directly from application code. It is used to define the
-    /// internals of the [`draw`] method used for the [`Circle`] primitive. To draw a circle, call
-    /// [`draw`] on a [`Circle`] primitive object.
-    ///
-    /// [`Circle`]: ./primitives/circle/struct.Circle.html
-    /// [`draw`]: ./trait.DrawTarget.html#method.draw
-    fn draw_circle(&mut self, item: &Styled<primitives::Circle, PrimitiveStyle<C>>) {
-        self.draw_iter(item);
-    }
-}
+pub use draw_target::DrawTarget;
